Add Snapper compability
Snapper does not create snapshots with the date in the name. The snapshots are only numbered. The date and some other information is stored in the info.xml next to the snapshot. This code reads the date from the info.xml file created by snapper.
For this plugin to work with the snapper structure you have to use something like "/data/.snapshots/%snapshot%/snapshot" as Snapshot format (where %snapshot% = 1,2,3,...).